requirements.txt
requests
feedparser
numpy
tqdm
python-dotenv
